requirements.txt
Flask==2.2.3
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
discord.py==2.2.2
twilio==7.16.5
openai==0.27.1